Fix formatting issues in pascals_triangle.cpp

+++ pascals_triangle.cpp
@@ -14,7 +14,7 @@
  * @param n a non-negative integer.
  * @return n! (with 0! defined as 1).
  */
-int factorial(int n) {
+long long factorial(int n) {
    long long result = 1;
    for (int i = 2; i <= n; i++) {
       result *= i;
